// src/app/model/OfertaLaboral/Oferta.ts
export class Oferta {
    id_oferta: number = 0;
    titulo: string = '';
    subtitulo: string = '';
    id_area: number = 0;
    id_pais: number = 0;
    id_estado: number = 0;
    id_ciudad: number = 0;
    id_jornada_laboral: number = 0;
    id_tipo_contrato: number = 0;
    salario: number = 0;
    fecha_contratacion: string = '';
    vacantes: number = 0;
    oferta_urgente: number = 0;
    descripcion: string = '';
    id_experiencia: number = 0;
    id_estudios: number = 0;
    licencia: string = '';
    dis_viajar: number = 0;
    dis_camb_res: number = 0;
    discapacidad: number = 0;
    estado: number = 0;

    area: string = '';
    pais: string = '';
    lugar_estado: string = '';
    ciudad: string = '';
    jornada: string = '';
    contrato: string = '';
    salario_n: string = '';
    experiencia: string = '';
    estudios: string = '';
    viajar: string = '';
    res: string = '';
    discapacidad_n: string = '';

    id_colaborador: number = 0;
    nombres: string = '';
    apellido_p: string = '';
    apellido_m: string = '';

    fecha_contratacion_n: string = '';
    ultima_fecha: string = '';
    ultima_fecha_n: string = '';
    fecha_creacion: string = '';
    fecha_creacion_n: string = '';

    string_CiuEst: string = '';
    estado_n: string = '';
    urgente_n: string = '';
    titulo_comp: string = '';

    postulados: number = 0;
    finalistas: number = 0;
    seleccionados: number = 0;

    idioma: number[] = [];
    nivel: number[] = [];
}