// src/app/model/OfertaLaboral/Atraccionusuario.ts
export class Atraccionusuario {
    id_usuario: number = 0;
    correo: string = '';
    nombres: string = '';
    apellido_p: string = '';
    apellido_m: string = '';
    titulo_curriculum: string = '';
    pais: string = '';
    estado: string = '';
    nom_loc: string = '';
    edad: number = -1;
    fecha_nacimiento: string = '';
    fecha_nacimiento_n: string = '';
    estado_usuario: string = '';
    nacionalidad: string = '';
    visto: number = -1;
    prefijo: string = '';
    telefono: string = '';
    estado_n: string = '';

    segment: number = 0;
    segment_n: string = '';
}